Внесены правки соотвественно с комментариями, установлен normalize
