+++ package.json
@@ -9,6 +9,7 @@
     "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore"
   },
   "dependencies": {
+    "normalize.css": "^8.0.1",
     "vue": "^3.2.47",
     "vue-router": "^4.1.6"
   },
