+++ src/App.vue
@@ -1,15 +1,12 @@
 <script setup>
-import { RouterView } from 'vue-router'
+import { RouterView } from 'vue-router';
+import 'normalize.css';
 </script>
 
 <template>
   <RouterView />
 </template>
 
-<style>
-
-* {
-    margin: 0;
-}
+<style lang="scss">
 
 </style>
